+++ src/dash-adapter.js
@@ -78,10 +78,10 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    * @type {Object}
    * @private
    */
-  _adapterEventsBindings: { [name: string]: Function } = {
-    [ShakaEvent.ERROR]: (event) => this._onError(event),
+  _adapterEventsBindings: {[name: string]: Function} = {
+    [ShakaEvent.ERROR]: event => this._onError(event),
     [ShakaEvent.ADAPTATION]: () => this._onAdaptation(),
-    [ShakaEvent.BUFFERING]: (event) => this._onBuffering(event),
+    [ShakaEvent.BUFFERING]: event => this._onBuffering(event),
     [EventType.WAITING]: () => this._onWaiting(),
     [EventType.PLAYING]: () => this._onPlaying()
   };
@@ -115,6 +115,13 @@ export class DashAdapter extends BaseMediaSourceAdapter {
   _playingSent: boolean = false;
 
   /**
+   * Video size update timer
+   * @type {null|number}
+   * @private
+   */
+  _videoSizeUpdateTimer: ?number = null;
+
+  /**
    * 3016 is the number of the video error at shaka, we already listens to it in the html5 class
    * @member {number} - VIDEO_ERROR_CODE
    * @type {number}
@@ -132,11 +139,45 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    * @static
    */
   static createAdapter(videoElement: HTMLVideoElement, source: PKMediaSourceObject, config: Object): IMediaSourceAdapter {
-    let dashConfig = {};
-    if (Utils.Object.hasPropertyPath(config, 'playback.options.html5.dash')) {
-      dashConfig = config.playback.options.html5.dash;
+    let adapterConfig: Object = Utils.Object.copyDeep(DefaultConfig);
+    if (Utils.Object.hasPropertyPath(config, 'playback.useNativeTextTrack')) {
+      adapterConfig.textTrackVisibile = Utils.Object.getPropertyPath(config, 'playback.useNativeTextTrack');
     }
-    return new this(videoElement, source, dashConfig);
+    if (Utils.Object.hasPropertyPath(config, 'sources.options')) {
+      const options = config.sources.options;
+      adapterConfig.forceRedirectExternalStreams = options.forceRedirectExternalStreams;
+      adapterConfig.redirectExternalStreamsHandler = options.redirectExternalStreamsHandler;
+      adapterConfig.redirectExternalStreamsTimeout = options.redirectExternalStreamsTimeout;
+    }
+    if (Utils.Object.hasPropertyPath(config, 'abr')) {
+      const abr = config.abr;
+      if (typeof abr.enabled === 'boolean') {
+        adapterConfig.shakaConfig.abr.enabled = abr.enabled;
+      }
+      if (typeof abr.capLevelToPlayerSize === 'boolean') {
+        adapterConfig.capLevelToPlayerSize = abr.capLevelToPlayerSize;
+      }
+      if (abr.defaultBandwidthEstimate) {
+        adapterConfig.shakaConfig.abr.defaultBandwidthEstimate = abr.defaultBandwidthEstimate;
+      }
+      if (abr.restrictions) {
+        if (abr.restrictions.minBitrate > 0) {
+          adapterConfig.shakaConfig.abr.restrictions.minBandwidth = abr.restrictions.minBitrate;
+        }
+        if (abr.restrictions.maxBitrate < Infinity) {
+          //You can either set capping by size or bitrate, if bitrate is set then disable size capping
+          adapterConfig.capLevelToPlayerSize = false;
+          adapterConfig.shakaConfig.abr.restrictions.maxBandwidth = abr.restrictions.maxBitrate;
+        }
+      }
+    }
+
+    //Merge shaka config with override config, override takes precedence
+    if (Utils.Object.hasPropertyPath(config, 'playback.options.html5.dash')) {
+      Utils.Object.mergeDeep(adapterConfig.shakaConfig, config.playback.options.html5.dash);
+    }
+
+    return new this(videoElement, source, adapterConfig);
   }
 
   /**
@@ -153,18 +194,59 @@ export class DashAdapter extends BaseMediaSourceAdapter {
   }
 
   /**
+   * set 'bitrate' the max bandwidth (if possible)
+   * @param {number} bitrate the max bitrate allowed
+   * @returns {void}
+   */
+  setMaxBitrate(bitrate: number): void {
+    if (this._hasLowerOrEqualBitrate(bitrate)) {
+      this._shaka.configure({abr: {restrictions: {maxBandwidth: bitrate}}});
+    }
+  }
+
+  _getSortedTracks(): Array<Object> {
+    const tracks = this._shaka.getVariantTracks();
+    const sortedTracks = tracks
+      .map(obj => ({
+        id: obj.id,
+        bandwidth: obj.bandwidth,
+        active: obj.active
+      }))
+      .sort((obj1, obj2) => obj1.bandwidth - obj2.bandwidth);
+    return sortedTracks;
+  }
+
+  _hasLowerOrEqualBitrate(maxBitrate: number): boolean {
+    const sortedTracks = this._getSortedTracks();
+    if (sortedTracks[0].bandwidth <= maxBitrate) {
+      return true;
+    }
+    return false;
+  }
+
+  /**
    * Checks if dash adapter can play a given drm data.
    * @param {Array<Object>} drmData - The drm data to check.
+   * @param {PKDrmConfigObject} drmConfig - The drm config.
    * @returns {boolean} - Whether the dash adapter can play a specific drm data.
    * @static
    */
-  static canPlayDrm(drmData: Array<Object>): boolean {
+  static canPlayDrm(drmData: Array<Object>, drmConfig: PKDrmConfigObject): boolean {
     let canPlayDrm = false;
     for (let drmProtocol of DashAdapter._drmProtocols) {
-      if (drmProtocol.canPlayDrm(drmData)) {
+      if (drmProtocol.isConfigured(drmData, drmConfig)) {
         DashAdapter._drmProtocol = drmProtocol;
         canPlayDrm = true;
         break;
+      }
+    }
+    if (!canPlayDrm) {
+      for (let drmProtocol of DashAdapter._drmProtocols) {
+        if (drmProtocol.canPlayDrm(drmData)) {
+          DashAdapter._drmProtocol = drmProtocol;
+          canPlayDrm = true;
+          break;
+        }
       }
     }
     DashAdapter._logger.debug('canPlayDrm result is ' + canPlayDrm.toString(), drmData);
@@ -216,11 +298,13 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    */
   _setShakaConfig(): void {
     const textDisplayerConfig = {
-      textDisplayFactory: function (videoEl) {
-        return new TextDisplayer(videoEl);
-      }.bind(null, this._videoElement)
+      shakaConfig: {
+        textDisplayFactory: function(videoEl) {
+          return new TextDisplayer(videoEl);
+        }.bind(null, this._videoElement)
+      }
     };
-    this._config = Utils.Object.mergeDeep(textDisplayerConfig, this._config, DefaultConfig);
+    this._config = Utils.Object.mergeDeep(textDisplayerConfig, DefaultConfig, this._config);
   }
 
   /**
@@ -233,9 +317,32 @@ export class DashAdapter extends BaseMediaSourceAdapter {
     shaka.polyfill.installAll();
     this._shaka = new shaka.Player(this._videoElement);
     this._maybeSetDrmConfig();
-    this._shaka.configure(this._config);
-    this._shaka.setTextTrackVisibility(false);
+    this._shaka.configure(this._config.shakaConfig);
     this._addBindings();
+  }
+
+  /**
+   * get the redirected URL
+   * @param {string} url - The url to check for redirection
+   * @returns {Promise<string>} - the resolved url
+   * @private
+   */
+  _maybeGetRedirectedUrl(url: string): Promise<string> {
+    const shouldRedirect = this._config.forceRedirectExternalStreams;
+    const timeout = this._config.redirectExternalStreamsTimeout;
+    const callback = this._config.redirectExternalStreamsHandler;
+    return new Promise(resolve => {
+      if (!shouldRedirect) {
+        return resolve(url);
+      }
+      Utils.Http.jsonp(url, callback, {
+        timeout: timeout
+      })
+        .then(uri => {
+          resolve(uri);
+        })
+        .catch(() => resolve(url));
+    });
   }
 
   /**
@@ -245,8 +352,97 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    */
   _maybeSetDrmConfig(): void {
     if (DashAdapter._drmProtocol && this._sourceObj && this._sourceObj.drmData) {
-      DashAdapter._drmProtocol.setDrmPlayback(this._config, this._sourceObj.drmData);
+      DashAdapter._drmProtocol.setDrmPlayback(this._config.shakaConfig, this._sourceObj.drmData);
     }
+  }
+
+  /**
+   * apply ABR restrictions
+   * @private
+   * @returns {void}
+   */
+  _maybeApplyAbrRestrictions(): void {
+    if (this._config.capLevelToPlayerSize) {
+      const videoTracks = this._getVideoTracks();
+      const getMinDimensions = (dim): number => Math.min.apply(null, videoTracks.map(variant => variant[dim]));
+      //Get minimal allowed dimensions
+      const minWidth = getMinDimensions('width');
+      const minHeight = getMinDimensions('height');
+      const updateAbrRestrictions = () => {
+        const curHeight = this._videoHeight;
+        const curWidth = this._videoWidth;
+        if (typeof curWidth === 'number' && typeof curHeight === 'number') {
+          //check if current player size is smaller than smallest rendition
+          //setting restriction below smallest rendition size will result in shaka emitting restriction unmet error
+          if (curHeight >= minHeight && curWidth >= minWidth) {
+            DashAdapter._logger.debug(`applying dimension restriction: width < ${curWidth}, height < ${curHeight}`);
+            this._shaka.configure({
+              abr: {
+                restrictions: {
+                  maxHeight: curHeight,
+                  maxWidth: curWidth
+                }
+              }
+            });
+          } else {
+            DashAdapter._logger.debug(`applying dimension restriction: width < ${minHeight}, height < ${minWidth}`);
+            this._shaka.configure({
+              abr: {
+                restrictions: {
+                  maxHeight: minHeight,
+                  maxWidth: minWidth
+                }
+              }
+            });
+          }
+        }
+      };
+      this._clearVideoUpdateTimer();
+      this._videoSizeUpdateTimer = setInterval(updateAbrRestrictions, ABR_RESTRICTION_UPDATE_INTERVAL);
+      updateAbrRestrictions();
+    }
+  }
+
+  /**
+   * Clear the video update timer
+   * @private
+   * @returns {void}
+   */
+  _clearVideoUpdateTimer(): void {
+    if (this._videoSizeUpdateTimer) {
+      clearInterval(this._videoSizeUpdateTimer);
+      this._videoSizeUpdateTimer = null;
+    }
+  }
+
+  get _videoWidth(): ?number {
+    let width;
+    const videoElement = this._videoElement;
+    if (videoElement) {
+      width = videoElement.width || videoElement.clientWidth || videoElement.offsetWidth;
+      width *= this._contentScaleFactor;
+    }
+    return width;
+  }
+
+  get _videoHeight(): ?number {
+    let height;
+    const videoElement = this._videoElement;
+    if (videoElement) {
+      height = videoElement.height || videoElement.clientHeight || videoElement.offsetHeight;
+      height *= this._contentScaleFactor;
+    }
+    return height;
+  }
+
+  get _contentScaleFactor(): number {
+    let pixelRatio = 1;
+    try {
+      pixelRatio = window.devicePixelRatio;
+    } catch (e) {
+      DashAdapter._logger.debug('failed reading devicePixelRatio, assume 1');
+    }
+    return pixelRatio;
   }
 
   /**
@@ -261,6 +457,17 @@ export class DashAdapter extends BaseMediaSourceAdapter {
     this._shaka.addEventListener(ShakaEvent.BUFFERING, this._adapterEventsBindings.buffering);
     this._videoElement.addEventListener(EventType.WAITING, this._adapterEventsBindings.waiting);
     this._videoElement.addEventListener(EventType.PLAYING, this._adapterEventsBindings.playing);
+    // called when a resource is downloaded
+    this._shaka.getNetworkingEngine().registerResponseFilter((type, response) => {
+      switch (type) {
+        case shaka.net.NetworkingEngine.RequestType.SEGMENT:
+          this._trigger(EventType.FRAG_LOADED, {miliSeconds: response.timeMs, bytes: response.data.byteLength});
+          break;
+        case shaka.net.NetworkingEngine.RequestType.MANIFEST:
+          this._trigger(EventType.MANIFEST_LOADED, {miliSeconds: response.timeMs});
+          break;
+      }
+    });
   }
 
   /**
@@ -289,18 +496,20 @@ export class DashAdapter extends BaseMediaSourceAdapter {
       this._loadPromise = new Promise((resolve, reject) => {
         if (this._sourceObj && this._sourceObj.url) {
           this._trigger(EventType.ABR_MODE_CHANGED, {mode: this.isAdaptiveBitrateEnabled() ? 'auto' : 'manual'});
-          const shakaStartTime = (startTime && startTime > -1 ? startTime : undefined);
-          this._shaka.load(this._sourceObj.url, shakaStartTime).then(() => {
-            let data = {tracks: this._getParsedTracks()};
-            DashAdapter._logger.debug('The source has been loaded successfully');
-            resolve(data);
-          }).catch((error) => {
-            reject(new Error(
-              error.severity,
-              error.category,
-              error.code,
-              error.data));
-          });
+          const shakaStartTime = startTime && startTime > -1 ? startTime : undefined;
+          this._maybeGetRedirectedUrl(this._sourceObj.url)
+            .then(url => {
+              return this._shaka.load(url, shakaStartTime);
+            })
+            .then(() => {
+              let data = {tracks: this._getParsedTracks()};
+              this._maybeApplyAbrRestrictions();
+              DashAdapter._logger.debug('The source has been loaded successfully');
+              resolve(data);
+            })
+            .catch(error => {
+              reject(new Error(error.severity, error.category, error.code, error.data));
+            });
         }
       });
     }
@@ -316,15 +525,16 @@ export class DashAdapter extends BaseMediaSourceAdapter {
   destroy(): Promise<*> {
     return super.destroy().then(() => {
       DashAdapter._logger.debug('destroy');
+      this._clearVideoUpdateTimer();
       this._loadPromise = null;
       this._buffering = false;
       this._waitingSent = false;
       this._playingSent = false;
       if (this._shaka) {
         this._removeBindings();
+        this._adapterEventsBindings = {};
         return this._shaka.destroy();
       }
-      this._adapterEventsBindings = {};
     });
   }
 
@@ -336,10 +546,10 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    */
   _getVideoTracks(): Array<Object> {
     let variantTracks = this._shaka.getVariantTracks();
-    let activeVariantTrack = variantTracks.filter((variantTrack) => {
+    let activeVariantTrack = variantTracks.filter(variantTrack => {
       return variantTrack.active;
     })[0];
-    let videoTracks = variantTracks.filter((variantTrack) => {
+    let videoTracks = variantTracks.filter(variantTrack => {
       return variantTrack.audioId === activeVariantTrack.audioId;
     });
     return videoTracks;
@@ -348,16 +558,19 @@ export class DashAdapter extends BaseMediaSourceAdapter {
   /**
    * Get the original audio tracks
    * @function _getAudioTracks
-   * @returns {Array<Object>} - The original audio tracks
+   * @returns {Array<Object>} - Array of objects with unique language and label.
    * @private
    */
   _getAudioTracks(): Array<Object> {
-    let variantTracks = this._shaka.getVariantTracks();
-    let activeVariantTrack = variantTracks.filter((variantTrack) => {
-      return variantTrack.active;
-    })[0];
-    let audioTracks = variantTracks.filter((variantTrack) => {
-      return variantTrack.videoId === activeVariantTrack.videoId;
+    const variantTracks = this._shaka.getVariantTracks();
+    let audioTracks = this._shaka.getAudioLanguagesAndRoles();
+    audioTracks.forEach(track => {
+      const sameLangAudioVariants = variantTracks.filter(vt => vt.language === track.language);
+      const id = sameLangAudioVariants.map(variant => variant.id).join('_');
+      const active = sameLangAudioVariants.some(variant => variant.active);
+      track.id = id;
+      track.label = sameLangAudioVariants[0].label;
+      track.active = active;
     });
     return audioTracks;
   }
@@ -439,7 +652,7 @@ export class DashAdapter extends BaseMediaSourceAdapter {
     if (textTracks) {
       for (let i = 0; i < textTracks.length; i++) {
         let settings = {
-          kind: textTracks[i].kind ? textTracks[i].kind + 's' : "",
+          kind: textTracks[i].kind ? textTracks[i].kind + 's' : '',
           active: false,
           label: textTracks[i].label,
           language: textTracks[i].language,
@@ -461,7 +674,7 @@ export class DashAdapter extends BaseMediaSourceAdapter {
   selectVideoTrack(videoTrack: VideoTrack): void {
     if (this._shaka) {
       let videoTracks = this._getVideoTracks();
-      if ((videoTrack instanceof VideoTrack) && videoTracks) {
+      if (videoTrack instanceof VideoTrack && videoTracks) {
         let selectedVideoTrack = videoTracks[videoTrack.index];
         if (selectedVideoTrack) {
           if (this.isAdaptiveBitrateEnabled()) {
@@ -485,7 +698,7 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    * @public
    */
   selectAudioTrack(audioTrack: AudioTrack): void {
-    if (this._shaka && (audioTrack instanceof AudioTrack) && !audioTrack.active) {
+    if (this._shaka && audioTrack instanceof AudioTrack && !audioTrack.active) {
       this._shaka.selectAudioLanguage(audioTrack.language);
       this._onTrackChanged(audioTrack);
     }
@@ -499,7 +712,8 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    * @public
    */
   selectTextTrack(textTrack: TextTrack): void {
-    if (this._shaka && (textTrack instanceof TextTrack) && !textTrack.active && (textTrack.kind === 'subtitles' || textTrack.kind === 'captions')) {
+    if (this._shaka && textTrack instanceof TextTrack && !textTrack.active && (textTrack.kind === 'subtitles' || textTrack.kind === 'captions')) {
+      this._shaka.setTextTrackVisibility(this._config.textTrackVisibile);
       this._shaka.selectTextLanguage(textTrack.language);
       this._onTrackChanged(textTrack);
     }
@@ -560,7 +774,7 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    * @public
    */
   seekToLiveEdge(): void {
-    if (this._shaka) {
+    if (this._shaka && this._videoElement.readyState > 0) {
       this._videoElement.currentTime = this._shaka.seekRange().end;
     }
   }
@@ -585,9 +799,10 @@ export class DashAdapter extends BaseMediaSourceAdapter {
    * @private
    */
   _onAdaptation(): void {
-    let selectedVideoTrack = this._getParsedVideoTracks().filter(function (videoTrack) {
+    let selectedVideoTrack = this._getParsedVideoTracks().filter(function(videoTrack) {
       return videoTrack.active;
     })[0];
+    DashAdapter._logger.debug('Video track changed', selectedVideoTrack);
     this._onTrackChanged(selectedVideoTrack);
   }
 
@@ -605,12 +820,7 @@ export class DashAdapter extends BaseMediaSourceAdapter {
       if (error.code === this.VIDEO_ERROR_CODE) {
         return;
       }
-      this._trigger(EventType.ERROR,
-        new Error(
-          error.severity,
-          error.category,
-          error.code,
-          error.data));
+      this._trigger(EventType.ERROR, new Error(error.severity, error.category, error.code, error.data));
       DashAdapter._logger.error(error);
     }
   }
@@ -631,7 +841,8 @@ export class DashAdapter extends BaseMediaSourceAdapter {
       }
     } else {
       this._buffering = false;
-      if (!this._videoElement.paused && !this._playingSent) { //the player leaves the buffering state. and 'playing' event hasn't been sent by the HTMLVideoElement.
+      if (!this._videoElement.paused && !this._playingSent) {
+        //the player leaves the buffering state. and 'playing' event hasn't been sent by the HTMLVideoElement.
         this._videoElement.dispatchEvent(new window.Event(EventType.PLAYING));
       }
     }
@@ -674,6 +885,29 @@ export class DashAdapter extends BaseMediaSourceAdapter {
     } else {
       return 0;
     }
+  }
+
+  /**
+   * gets the target buffer of the player
+   * @returns {number} - buffer length target in seconds
+   */
+  get targetBuffer(): number {
+    let targetBufferVal = NaN;
+    if (!this._shaka) return NaN;
+    if (this.isLive()) {
+      if (this._shaka.getManifest()) {
+        targetBufferVal =
+          this._shaka.getManifest().presentationTimeline.getSegmentAvailabilityEnd() -
+          this._shaka.getManifest().presentationTimeline.getSeekRangeEnd() -
+          (this._videoElement.currentTime - this._getLiveEdge());
+        targetBufferVal = Math.min(targetBufferVal, this._shaka.getConfiguration().streaming.bufferingGoal);
+      }
+    } else {
+      if (this._shaka.getConfiguration() && this._shaka.getConfiguration().streaming) {
+        targetBufferVal = this._shaka.getConfiguration().streaming.bufferingGoal;
+      }
+    }
+    return targetBufferVal;
   }
 }